# SegmentFromPoint: Speed Up (EfficientSAM) & Return Feedback