--- a/ada_feeding_perception/ada_feeding_perception/segment_from_point.py
+++ b/ada_feeding_perception/ada_feeding_perception/segment_from_point.py
@@ -5,15 +5,14 @@ server that takes in a seed point, segments the latest image with that seed
 point using Segment Anything, and returns the top n contender masks.
 """
 # Standard imports
-import math
 import os
-import reprlib
 import threading
 from typing import Optional, Tuple, Union
 
 # Third-party imports
 import cv2
 from cv_bridge import CvBridge
+from efficient_sam.efficient_sam import build_efficient_sam
 import numpy as np
 import numpy.typing as npt
 from rcl_interfaces.msg import ParameterDescriptor, ParameterType
@@ -71,6 +70,7 @@ class SegmentFromPointNode(Node):
             model_name,
             model_base_url,
             model_dir,
+            self.use_efficient_sam,
             self.n_contender_masks,
         ) = self.read_params()
 
@@ -140,7 +140,24 @@ class SegmentFromPointNode(Node):
         )
 
         # Initialize Segment Anything
-        self.initialize_food_segmentation(model_name, model_path)
+        if self.use_efficient_sam:
+            self.initialize_efficient_sam(model_name, model_path)
+        else:
+            self.initialize_sam(model_name, model_path)
+
+        # Convert between ROS and CV images
+        self.bridge = CvBridge()
+
+        # Create the Action Server
+        self._action_server = ActionServer(
+            self,
+            SegmentFromPoint,
+            "SegmentFromPoint",
+            self.execute_callback,
+            goal_callback=self.goal_callback,
+            cancel_callback=self.cancel_callback,
+            callback_group=MutuallyExclusiveCallbackGroup(),
+        )
 
     def read_params(
         self,
@@ -154,12 +171,14 @@ class SegmentFromPointNode(Node):
         model_base_url: The URL to download the model checkpoint from if it is
             not already downloaded
         model_dir: The location of the directory where the model checkpoint is / should be stored
+        use_efficient_sam: Whether to use EfficientSAM or SAM
         n_contender_masks: The number of contender masks to return per point.
         """
         (
             model_name,
             model_base_url,
             model_dir,
+            use_efficient_sam,
             n_contender_masks,
         ) = self.declare_parameters(
             "",
@@ -170,7 +189,7 @@ class SegmentFromPointNode(Node):
                     ParameterDescriptor(
                         name="model_name",
                         type=ParameterType.PARAMETER_STRING,
-                        description="The name of the Segment Anything model checkpoint to use",
+                        description="The name of the model checkpoint to use",
                         read_only=True,
                     ),
                 ),
@@ -201,6 +220,16 @@ class SegmentFromPointNode(Node):
                     ),
                 ),
                 (
+                    "use_efficient_sam",
+                    True,
+                    ParameterDescriptor(
+                        name="use_efficient_sam",
+                        type=ParameterType.PARAMETER_BOOL,
+                        description=("Whether to use EfficientSAM or SAM"),
+                        read_only=True,
+                    ),
+                ),
+                (
                     "n_contender_masks",
                     3,
                     ParameterDescriptor(
@@ -216,14 +245,15 @@ class SegmentFromPointNode(Node):
             model_name.value,
             model_base_url.value,
             model_dir.value,
+            use_efficient_sam.value,
             n_contender_masks.value,
         )
 
-    def initialize_food_segmentation(self, model_name: str, model_path: str) -> None:
+    def initialize_sam(self, model_name: str, model_path: str) -> None:
         """
-        Initialize all attributes needed for food segmentation.
+        Initialize all attributes needed for food segmentation with SAM.
 
-        This includes loading the Segment Anything model, launching the action
+        This includes loading the SAM, launching the action
         server, and more. Note that we are guarenteed the model exists since
         it was downloaded in the __init__ function of this class.
 
@@ -236,7 +266,7 @@ class SegmentFromPointNode(Node):
         ------
         ValueError if the model name does not contain vit_h, vit_l, or vit_b
         """
-        self.get_logger().info("Initializing food segmentation...")
+        self.get_logger().info("Initializing SAM...")
         # Load the model and move it to the specified device
         if "vit_b" in model_name:  # base model
             model_type = "vit_b"
@@ -254,19 +284,42 @@ class SegmentFromPointNode(Node):
         # a lock.
         self.predictor = SamPredictor(sam)
 
-        # Convert between ROS and CV images
-        self.bridge = CvBridge()
+        self.get_logger().info("...Done!")
 
-        # Create the action server
-        self._action_server = ActionServer(
-            self,
-            SegmentFromPoint,
-            "SegmentFromPoint",
-            self.execute_callback,
-            goal_callback=self.goal_callback,
-            cancel_callback=self.cancel_callback,
-            callback_group=MutuallyExclusiveCallbackGroup(),
-        )
+    def initialize_efficient_sam(self, model_name: str, model_path: str) -> None:
+        """
+        Initialize all attributes needed for food segmentation with EfficientSAM.
+
+        This includes loading the EfficientSAM model, launching the action
+        server, and more. Note that we are guarenteed the model exists since
+        it was downloaded in the __init__ function of this class.
+
+        Parameters
+        ----------
+        model_name: The name of the model to load.
+        model_path: The path to the model checkpoint to load.
+
+        Raises
+        ------
+        ValueError if the model name does not contain efficient_sam
+        """
+        self.get_logger().info("Initializing EfficientSAM...")
+        # Hardcoded from https://github.com/yformer/EfficientSAM/blob/main/efficient_sam/build_efficient_sam.py
+        if "vits" in model_name:
+            encoder_patch_embed_dim = 384
+            encoder_num_heads = 6
+        elif "vitt" in model_name:
+            encoder_patch_embed_dim = 192
+            encoder_num_heads = 3
+        else:
+            raise ValueError(f"Unknown model type {model_name}")
+        self.efficient_sam = build_efficient_sam(
+            encoder_patch_embed_dim=encoder_patch_embed_dim,
+            encoder_num_heads=encoder_num_heads,
+            checkpoint=model_path,
+        ).eval()
+        self.efficient_sam.to(device=self.device)
+
         self.get_logger().info("...Done!")
 
     def camera_info_callback(self, msg: CameraInfo) -> None:
@@ -351,6 +404,83 @@ class SegmentFromPointNode(Node):
         self.get_logger().info("Received cancel request, accepting")
         return CancelResponse.ACCEPT
 
+    def run_sam(
+        self, image: npt.NDArray, seed_point: Tuple[int, int]
+    ) -> list[Tuple[float, npt.NDArray[np.bool_]]]:
+        """
+        Run the SAM model on the image and seed point.
+
+        Parameters
+        ----------
+        image: The image to segment, in BGR.
+        seed_point: The seed point to segment from.
+
+        Returns
+        -------
+        masks: The masks outputted by the model.
+        scores: The scores outputted by the model.
+        """
+        self.get_logger().info("Segmenting image with SAM...")
+        # Convert image from BGR to RGB
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
+
+        # Segment the image with that input point
+        self.predictor.set_image(image)
+        masks, scores, _ = self.predictor.predict(
+            point_coords=np.array([seed_point]),
+            point_labels=np.array([1]),
+            multimask_output=True,
+        )
+        return masks, scores
+
+    def run_efficient_sam(
+        self, image: npt.NDArray, seed_point: Tuple[int, int]
+    ) -> list[Tuple[float, npt.NDArray[np.bool_]]]:
+        """
+        Run the EfficientSAM model on the image and seed point.
+
+        Code adapted from https://github.com/yformer/EfficientSAM/blob/main/EfficientSAM_example.py
+
+        Parameters
+        ----------
+        image: The image to segment, in BGR.
+        seed_point: The seed point to segment from.
+
+        Returns
+        -------
+        masks: The masks outputted by the model.
+        scores: The scores outputted by the model.
+        """
+        self.get_logger().info("Segmenting image with EfficientSAM...")
+        # Convert image from BGR to RGB
+        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
+
+        # Convert the image to a tensor
+        image_tensor = torch.tensor(image).to(device=self.device)
+
+        # Convert the seed point to a tensor
+        input_points = torch.tensor(np.reshape(seed_point, (1, 1, 1, 2))).to(
+            device=self.device
+        )
+
+        # Convert the labels to a tensor
+        input_labels = torch.tensor([[[1]]]).to(device=self.device)
+
+        # Run the model
+        predicted_logits, predicted_iou = self.efficient_sam(
+            image_tensor[None, ...],
+            input_points,
+            input_labels,
+        )
+        sorted_ids = torch.argsort(predicted_iou, dim=-1, descending=True)
+        predicted_iou = torch.take_along_dim(predicted_iou, sorted_ids, dim=2)
+        predicted_logits = torch.take_along_dim(
+            predicted_logits, sorted_ids[..., None, None], dim=2
+        )
+        masks = torch.ge(predicted_logits[0, 0, :, :, :], 0).cpu().detach().numpy()
+        scores = predicted_iou[0, 0, :].cpu().detach().numpy()
+        return masks, scores
+
     def segment_image(
         self, seed_point: Tuple[int, int], image_msg: Image
     ) -> SegmentFromPoint.Result:
@@ -385,23 +515,15 @@ class SegmentFromPointNode(Node):
         # Convert the image to OpenCV format
         image = ros_msg_to_cv2_image(image_msg, self.bridge)
 
-        # Convert the image to RGB for Segment Anything
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-
         # Convert the depth image to OpenCV format. The depth image is a
         # 16-bit image with depth in mm.
         depth_img = ros_msg_to_cv2_image(depth_img_msg, self.bridge)
 
         # Segment the image
-        self.predictor.set_image(image)
-        masks, scores, _ = self.predictor.predict(
-            point_coords=np.array([seed_point]),
-            point_labels=np.array([1]),
-            multimask_output=True,
-        )
-
-        # Convert the image back to BGR, which is OpenCV's default format
-        image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
+        if self.use_efficient_sam:
+            masks, scores = self.run_efficient_sam(image, seed_point)
+        else:
+            masks, scores = self.run_sam(image, seed_point)
 
         # Sort the masks from highest to lowest score
         scored_masks_sorted = sorted(
@@ -421,9 +543,6 @@ class SegmentFromPointNode(Node):
         # Return the result message
         return result
 
-    # pylint: disable=too-many-arguments
-    # More arguments is fine here, since the purpose of this function is to consolidate
-    # multiple pieces of data into a single message.
     def generate_mask_msg(
         self,
         item_id: str,
@@ -445,6 +564,10 @@ class SegmentFromPointNode(Node):
         depth_img: The most recent depth image.
         seed_point: The seed point used to segment the image.
         """
+        # pylint: disable=too-many-arguments, too-many-locals
+        # More arguments is fine here, since the purpose of this function is to consolidate
+        # multiple pieces of data into a single message.
+
         # Clean the mask to only contain the connected component containing
         # the seed point
         # TODO: Thoroughly test this, in case we need to add more mask cleaning!