--- a/ada_feeding_perception/config/segment_from_point.yaml
+++ b/ada_feeding_perception/config/segment_from_point.yaml
@@ -1,9 +1,15 @@
 # NOTE: You have to change this node name if you change the node name in the launchfile.
 segment_from_point:
   ros__parameters:
-    # The name of the Segment Anything model checkpoint to use
-    model_name: sam_vit_b_01ec64.pth
+    # # The name of the Segment Anything model checkpoint to use
+    # model_name: sam_vit_b_01ec64.pth
+    # # The URL to download the model checkpoint from if it is not already downloaded
+    # model_base_url: "https://dl.fbaipublicfiles.com/segment_anything/"
+
+    # The name of the Efficient Segment Anything model checkpoint to use
+    model_name: efficient_sam_vitt.pt
     # The URL to download the model checkpoint from if it is not already downloaded
-    model_base_url: "https://dl.fbaipublicfiles.com/segment_anything/"
+    model_base_url: "https://raw.githubusercontent.com/yformer/EfficientSAM/main/weights/"
+
     # The number of contender masks to return
     n_contender_masks: 3