--- a/ada_feeding_perception/config/segment_from_point.yaml
+++ b/ada_feeding_perception/config/segment_from_point.yaml
@@ -5,11 +5,13 @@ segment_from_point:
     # model_name: sam_vit_b_01ec64.pth
     # # The URL to download the model checkpoint from if it is not already downloaded
     # model_base_url: "https://dl.fbaipublicfiles.com/segment_anything/"
+    # use_efficient_sam: false
 
     # The name of the Efficient Segment Anything model checkpoint to use
     model_name: efficient_sam_vitt.pt
     # The URL to download the model checkpoint from if it is not already downloaded
     model_base_url: "https://raw.githubusercontent.com/yformer/EfficientSAM/main/weights/"
+    use_efficient_sam: true
 
     # The number of contender masks to return
     n_contender_masks: 3