… ada_feeding_perception/ada_feeding_perception/test_efficient_sam.py